requirements.txt
Updated requirements.txt with authentication and analytics dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.19
Alembic==1.12.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.4
requests==2.31.0
urllib3==2.0.4
anthropic>=0.69.0
python-dotenv==1.0.0
redis==5.0.0
Flask-Caching==2.1.0
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dateutil==2.8.2
pytest==7.4.2
pytest-flask==1.2.1
pytest-cov==4.1.0
gunicorn==21.2.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
click==8.1.7
marshmallow==3.20.1
